fixup! Swap to raw requests with caching
